I2C - Ensure Stop is send correctly
The stop bit was being requested but without waiting for the hardware to complete it (as no interrupt is generated).  At normal clock speeds the latency in the code is enough time for the stop bit to be completed, but at slow clock speeds it is necessary to wait for the stop bit to be sent.
Also, change low-level I2C routines, including IRC handler, to static functions to reduce overheads.