Reduce FLASH requirement for Nano/Uno
When IO_MINIMALHAL is defined, the functions that require implementation of specific HAL devices (such as I2C servos) are excluded from the build.  Also any code relating to Servos and Turnouts that rely on such devices are excluded, together with the command processing for them.  In this way, the Nano/Uno build fully supports I/O to Arduino pins but reduces the overall FLASH requirement.