Delay <H> responses for servo turnouts, until the servo movement is complete.
To allow JMRI to display a temporary state while a servo turnout is in motion, the <H> confirmation responses (when throwing/closing a turnout) are delayed until the readback from the servo pin indicates that the motion has completed.
NOTE: This does not apply to any other type of turnout, including VPIN turnouts.  If a VPIN turnout is polled for readback, it could force the pin into READ mode, which would be undesirable.